Support the logic for deleting contests
